Move segmentation application object into init
